ConDec-294: Move test classes
