+++ src/test/java/de/uhd/ifi/se/decision/management/jira/extraction/persistence/TestActiveObjectsManager.java
@@ -1,4 +1,4 @@
-package de.uhd.ifi.se.decision.management.jira.extraction.persistance;
+package de.uhd.ifi.se.decision.management.jira.extraction.persistence;
 
 import static org.junit.Assert.*;
